+++ release_notes.md
@@ -8,6 +8,7 @@ PREESM Changelog
 
 ### Changes
 *  Fix C Hardware codegen: properly prefix LOOP_SIZE with PREESM_;
+*  Update workflow editor to better report task id issues;
 
 ### Bug fix